made hadoopSeecurityUils Class
