+++ common/network-common/src/main/java/org/apache/spark/network/sasl/SparkSaslServer.java
@@ -27,8 +27,6 @@ import javax.security.sasl.RealmCallback;
 import javax.security.sasl.Sasl;
 import javax.security.sasl.SaslException;
 import javax.security.sasl.SaslServer;
-import java.io.ByteArrayInputStream;
-import java.io.DataInputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
@@ -42,9 +40,12 @@ import io.netty.handler.codec.base64.Base64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.hadoop.security.token.SecretManager.InvalidToken;
 import org.apache.hadoop.yarn.security.client.ClientToAMTokenSecretManager;
 import org.apache.hadoop.yarn.security.client.ClientToAMTokenIdentifier;
+
+import static org.apache.spark.network.util.HadoopSecurityUtils.decodeMasterKey;
+import static org.apache.spark.network.util.HadoopSecurityUtils.getClientToAMSecretKey;
+import static org.apache.spark.network.util.HadoopSecurityUtils.getIdentifier;
 
 import org.apache.spark.network.util.TransportConf;
 
@@ -223,30 +224,6 @@ public class SparkSaslServer implements SaslEncryptionBackend {
       }
     }
   }
-
-  /** Creates an ClientToAMTokenIdentifier from the encoded Base-64 String */
-  public static ClientToAMTokenIdentifier getIdentifier(String id) throws InvalidToken {
-    byte[] tokenId = byteBufToByte(Base64.decode(
-      Unpooled.wrappedBuffer(id.getBytes(StandardCharsets.UTF_8))));
-
-    ClientToAMTokenIdentifier tokenIdentifier = new ClientToAMTokenIdentifier();
-    try {
-      tokenIdentifier.readFields(new DataInputStream(new ByteArrayInputStream(tokenId)));
-    } catch (IOException e) {
-      throw (InvalidToken) new InvalidToken(
-              "Can't de-serialize tokenIdentifier").initCause(e);
-    }
-    return tokenIdentifier;
-  }
-
-  /** Returns an Base64-encoded secretKey created from the Identifier and the secretmanager */
-  public static char[] getClientToAMSecretKey(ClientToAMTokenIdentifier tokenid,
-                                        ClientToAMTokenSecretManager secretManager) throws InvalidToken {
-    byte[] password =  secretManager.retrievePassword(tokenid);
-    return Base64.encode(Unpooled.wrappedBuffer(password)).toString(StandardCharsets.UTF_8)
-            .toCharArray();
-  }
-
   /** Encode a String identifier as a Base64-encoded string. */
   public static String encodeIdentifier(String identifier) {
     Preconditions.checkNotNull(identifier, "User cannot be null if SASL is enabled");
@@ -259,25 +236,6 @@ public class SparkSaslServer implements SaslEncryptionBackend {
     return getBase64EncodedString(password).toCharArray();
   }
 
-  /** Decode a base64-encoded indentifier as a String. */
-  public static String decodeIdentifier(String identifier) {
-    Preconditions.checkNotNull(identifier, "User cannot be null if SASL is enabled");
-    return Base64.decode(Unpooled.wrappedBuffer(identifier.getBytes(StandardCharsets.UTF_8)))
-            .toString(StandardCharsets.UTF_8);
-  }
-
-  /** Decode a base64-encoded MasterKey as a byte[] array. */
-  public static byte[] decodeMasterKey(String masterKey) {
-    ByteBuf masterKeyByteBuf = Base64.decode(Unpooled.wrappedBuffer(masterKey.getBytes(StandardCharsets.UTF_8)));
-    return byteBufToByte(masterKeyByteBuf);
-  }
-
-  /** Convert an ByteBuf to a byte[] array. */
-  private static byte[] byteBufToByte(ByteBuf byteBuf) {
-    byte[] byteArray = new byte[byteBuf.readableBytes()];
-    byteBuf.readBytes(byteArray);
-    return byteArray;
-  }
   /** Return a Base64-encoded string. */
   private static String getBase64EncodedString(String str) {
     ByteBuf byteBuf = null;
